requirements.txt
flask
bulma
markdown
pyyaml